Progress on rendering single-material models, needs debugging and cleanup.
